ZioWriters/requirements.txt
Django>=5.2
djangorestframework
djangorestframework-authtoken
python-dotenv
psycopg2-binary   # only if you use PostgreSQL
